backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
drf-yasg
gunicorn
Pillow
django-cors-headers
dotenv